refactor: move src CSS to styles folder, use lit imports
