+++ packages/avatar/package.json
@@ -21,8 +21,8 @@
   "type": "module",
   "files": [
     "src",
-    "!src/vaadin-avatar-base-styles.d.ts",
-    "!src/vaadin-avatar-base-styles.js",
+    "!src/styles/*-base-styles.d.ts",
+    "!src/styles/*-base-styles.js",
     "theme",
     "vaadin-*.d.ts",
     "vaadin-*.js",
