+++ packages/avatar/src/styles/vaadin-avatar-base-styles.d.ts
