+++ packages/avatar/src/styles/vaadin-avatar-core-styles.d.ts
