+++ packages/avatar/src/styles/vaadin-avatar-core-styles.js
@@ -3,7 +3,7 @@
  * Copyright (c) 2020 - 2025 Vaadin Ltd.
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
-import { css } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { css } from 'lit';
 
 export const avatarStyles = css`
   :host {
